initial plumbing
