+++ app/controllers/SwitchesController.scala
@@ -8,5 +8,5 @@ import zio.DefaultRuntime
 import scala.concurrent.ExecutionContext
 
 class SwitchesController(authAction: AuthAction[AnyContent], components: ControllerComponents, stage: String, runtime: DefaultRuntime)(implicit ec: ExecutionContext)
-  extends S3ObjectController[SupportFrontendSwitches](authAction, components, stage, filename = "switches.json", runtime) {
+  extends S3ObjectController[SupportFrontendSwitches](authAction, components, stage, filename = "switches_v2.json", runtime) {
 }
